hp_printer_management/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
pysnmp==4.4.12
requests==2.31.0
Pillow==10.1.0
django-filter==23.4
drf-spectacular==0.26.5
django-extensions==3.2.3
python-decouple==3.8
django-import-export==3.3.1
openpyxl==3.1.2
reportlab==4.0.7
django-oauth-toolkit==1.7.1
ldap3==2.9.1
django-auth-ldap==4.6.0
python-ldap==3.4.3
smtplib2==0.2.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-environ==0.11.2
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest-django==4.7.0
coverage==7.3.2
flake8==6.1.0
black==23.11.0
django-silk==5.0.4
